refactor: agent form page to sepearte components
