+++ frontend/src/components/agent/GeneralTab.tsx
@@ -1,0 +1,172 @@
+import { FormField, FormItem, FormLabel, FormControl, FormDescription, FormMessage } from '@/components/ui/form';
+import { Input } from '@/components/ui/input';
+import { Textarea } from '@/components/ui/textarea';
+import { Slider } from '@/components/ui/slider';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { UseFormReturn } from 'react-hook-form';
+import type { AIProvider, AIModel } from '@/types/agent.types';
+import type { AgentFormValues } from './types';
+
+interface GeneralTabProps {
+  form: UseFormReturn<AgentFormValues>;
+  providers: AIProvider[];
+  models: AIModel[];
+  watchProvider: string;
+}
+
+export function GeneralTab({ form, providers, models, watchProvider }: GeneralTabProps) {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>LLM Configuration</CardTitle>
+        <CardDescription>Configure language model settings</CardDescription>
+      </CardHeader>
+      <CardContent className="grid gap-6 sm:grid-cols-2">
+        <FormField
+          control={form.control}
+          name="agent_name"
+          render={({ field }) => (
+            <FormItem className="sm:col-span-2">
+              <FormLabel>Agent Name</FormLabel>
+              <FormControl>
+                <Input placeholder="my-agent" {...field} />
+              </FormControl>
+              <FormDescription>Unique agent name</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="description"
+          render={({ field }) => (
+            <FormItem className="sm:col-span-2">
+              <FormLabel>Description</FormLabel>
+              <FormControl>
+                <Textarea
+                  placeholder="A short summary describing what this agent does or is designed for."
+                  className="min-h-[80px] resize-y"
+                  {...field}
+                />
+              </FormControl>
+              <FormDescription>A brief description of the agent's purpose</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="provider"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Provider</FormLabel>
+              <Select
+                onValueChange={(value) => {
+                  field.onChange(value);
+                  form.setValue('model', '');
+                }}
+                value={field.value}
+              >
+                <FormControl>
+                  <SelectTrigger>
+                    <SelectValue placeholder="Select provider" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  {providers.map((provider) => (
+                    <SelectItem key={provider.name} value={provider.name}>
+                      {provider.provider_name || provider.name}
+                    </SelectItem>
+                  ))}
+                </SelectContent>
+              </Select>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="model"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Model</FormLabel>
+              <Select
+                onValueChange={field.onChange}
+                value={field.value}
+                disabled={!watchProvider}
+              >
+                <FormControl>
+                  <SelectTrigger>
+                    <SelectValue placeholder="Select model" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  {models
+                    .filter((model) => model.provider === watchProvider)
+                    .map((model) => (
+                      <SelectItem key={model.name} value={model.name}>
+                        {model.model_name || model.name}
+                      </SelectItem>
+                    ))}
+                </SelectContent>
+              </Select>
+              <FormDescription>Filtered by selected provider</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="temperature"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Temperature: {field.value}</FormLabel>
+              <FormControl>
+                <Slider
+                  min={0}
+                  max={2}
+                  step={0.1}
+                  value={[field.value]}
+                  onValueChange={(vals) => field.onChange(vals[0])}
+                />
+              </FormControl>
+              <FormDescription>
+                Lower = focused, higher = creative
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="top_p"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Top P: {field.value}</FormLabel>
+              <FormControl>
+                <Slider
+                  min={0}
+                  max={1}
+                  step={0.05}
+                  value={[field.value]}
+                  onValueChange={(vals) => field.onChange(vals[0])}
+                />
+              </FormControl>
+              <FormDescription>
+                Nucleus sampling parameter
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+      </CardContent>
+    </Card>
+  );
+}
+
