+++ frontend/src/components/agent/types.ts
@@ -1,0 +1,17 @@
+import * as z from 'zod';
+
+export const agentFormSchema = z.object({
+  agent_name: z.string().min(1, 'Agent name is required'),
+  provider: z.string().min(1, 'Provider is required'),
+  model: z.string().min(1, 'Model is required'),
+  temperature: z.number().min(0).max(2),
+  top_p: z.number().min(0).max(1),
+  disabled: z.boolean(),
+  allow_chat: z.boolean(),
+  persist_conversation: z.boolean(),
+  description: z.string().optional(),
+  instructions: z.string(),
+});
+
+export type AgentFormValues = z.infer<typeof agentFormSchema>;
+
